feat(label): Add option to show/hide label count and stock info
This commit introduces a new configuration option to control the visibility of the "Label X of Y" and "Stock: Z" information on printed product labels.

A new boolean field, 'label_show_stock_label', has been added to the ResConfigSettings model and its corresponding view.

The report template has been updated to conditionally render this information based on the new setting.